feat: 공간 쿼리를 사용하기 위해 카페 [등록/수정] 수정 - 공간 쿼리를 사용하기 위해 카페 [등록/수정] 수정 - 주소를 위경도 좌표로 받아올 수 있는 mapService 추가 - Cafe entity location추가 - 테스트 코드 수정

+++ src/main/java/com/sparta/kidscafe/domain/cafe/dto/request/CafeSearchRequestDto.java
@@ -29,8 +29,8 @@ public class CafeSearchRequestDto extends CafeSearchPageRequestDto {
   private String openedAt;
   private String closedAt;
   private Long userId;
-  private Long lon;
-  private Long lat;
+  private Double lon;
+  private Double lat;
   private Double radiusMeter;
 
   public CafeSearchCondition getSearchCondition() {
